switch from libevent to a smaller, thin wrapper around poll()
libevent is a very cool library, I like its APIs and enjoy using
it.  However, for amused it is maybe a bit too much since it doesn't
deal with too many file descriptors.  Amused only needs to monitor
one fd for the player process and the currently connected clients.
Given that it runs per-user locally, having more than a dozen of
fds open is very, very rare.

+++ control.h
@@ -18,7 +18,7 @@
 
 int	control_init(char *);
 int	control_listen(int fd);
-void	control_accept(int, short, void *);
+void	control_accept(int, int, void *);
 void	control_notify(int);
-void	control_dispatch_imsg(int, short, void *);
+void	control_dispatch_imsg(int, int, void *);
 int	control_imsg_relay(struct imsg *);
